Extend Ruboty's classes using `Module#include` instead of replacing.
