--- lib/ruboty/robot.rb
@@ -1,10 +1,0 @@
-module Ruboty
-  class Robot
-    delegate :add_reaction, to: :adapter
-
-    def add_reaction(reaction, channel_id, timestamp)
-      adapter.add_reaction(reaction, channel_id, timestamp)
-      true
-    end
-  end
-end
